### 2022-08-08 #4661 @danielskatz
👋 @klb2 - I'm suggesting some minor grammar (and style) changes in https://github.com/klb2/qmkpy/pull/1. If you merge this, then please use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
