### 2022-08-09 #4661 @klb2
> wave @klb2 - I'm suggesting some minor grammar (and style) changes in [klb2/qmkpy#1](https://github.com/klb2/qmkpy/pull/1). If you merge this, then please use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Thanks a lot for the suggested changes. I merged them into the `joss-paper` branch.
