### 2022-08-13 2ee23a0
Add JOSS submission badge to README
