### 2022-08-13 c7adc60
Add JSON strategy for saving and loading problems
Add a new "json" IO strategy that allows saving and loading problems in
the JavaScript Object Notation (JSON) format.
