+++ qmkpy/qmkp.py
@@ -65,7 +65,8 @@ class QMKProblem:
         self.algorithm = algorithm
         self.args = args
 
-        self.assignments = np.zeros((len(self.weights), len(self.capacities)))
+        if assignments is None:
+            self.assignments = np.zeros((len(self.weights), len(self.capacities)))
     
     def solve(self, algorithm: Optional[Callable] = None,
               args: Optional[tuple] = None) -> Tuple[np.array, float]:
@@ -146,6 +147,7 @@ class QMKProblem:
             - ``txt``: Save the arrays of the model using the text-based format
               established by Billionnet and Soutif. See also
               :meth:`qmkpy.io.save_problem_txt()`.
+            - ``json``: Save the arrays of the model using the JSON format.
 
         Returns
         -------
@@ -159,6 +161,8 @@ class QMKProblem:
             io.save_problem_pickle(fname, self)
         elif strategy == "txt":
             io.save_problem_txt(fname, self)
+        elif strategy == "json":
+            io.save_problem_json(fname, self)
         else:
             raise NotImplementedError("The strategy '%s' is not implemented.",
                                       strategy)
@@ -192,6 +196,7 @@ class QMKProblem:
               module
             - ``txt``: Save the arrays of the model using the text-based format
               established by Billionnet and Soutif.
+            - ``json``: Save the arrays of the model using the JSON format.
 
         Returns
         -------
@@ -206,6 +211,8 @@ class QMKProblem:
             problem = io.load_problem_pickle(fname)
         elif strategy == "txt":
             problem = io.load_problem_txt(fname)
+        elif strategy == "json":
+            problem = io.load_problem_json(fname)
         else:
             raise NotImplementedError("The strategy '%s' is not implemented.",
                                       strategy)
