### 2022-08-13 0f11a9f
Add new name attribute to QMKProblem class
Add a new (optional) name attribute to the QMKProblem class that can
specify a name.
This name will be used in the text-based save strategies, if no other
name is provided.
