### 2022-08-22 #4661 @klb2
Thank you @jbytecode for serving as editor for this submission.

From the list of reviewers, I would suggest the following two reviewers:
- kramea
- ulf1

Both listed operations research as topic areas and Python as preferred programming language.
