### 2022-08-24 9b2f4a5
Add custom __str__ method to QMKProblem class
Add a new custom __str__ method to the qmkp.QMKProblem class which will
print the name attribute of the object, if it exists. Otherwise, it will
print a string that contains the number of items and knapsacks.
