### 2022-08-26 #4661 @ulf1
Hello @jbytecode , is there a checklist or criteria list I would need to apply for the review?
