### 2022-08-29 #4661 @max-little
Yes that's fine!
Best
Max

On Mon, 29 Aug 2022 at 18:02, Mehmet Hakan Satman ***@***.***>
wrote:

> 👋👋👋 Dear @max-little <https://github.com/max-little> 👋👋👋
>
> Would you be willing to assist in reviewing this submission for *JOSS* (*Journal
> of Open Source Software*)?
>
> *We need a 2nd reviewers for this submisson*
>
> JOSS publishes articles about open source research software. The
> submission I'd like you to review is titled: "*QMKPy: A Python Framework
> for the Quadratic Multiple Knapsack Problem*". You can find more
> information at the top of this Github issue (#4661
> <https://github.com/openjournals/joss-reviews/issues/4661>).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged. If you have
> any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will
> start the review process in a seperate thread. In that thread, there will
> be 20 check items for each single reviewer.
>
> Thank you in advance!
>
