### 2022-08-29 #4718 @editorialbot
[REVIEW]: QMKPy: A Python Framework for the Quadratic Multiple Knapsack Problem
**Submitting author:** @klb2 (<a href="http://orcid.org/0000-0002-1597-8963">Karl-Ludwig Besser</a>)
**Repository:** https://github.com/klb2/qmkpy
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.1.0
