### 2022-09-01 #4718 @ulf1
@klb2 May I ask what the second author Eduard A. Jorswieck contributed exactly? He does not seem to made *software contributions* (see https://github.com/klb2/qmkpy/graphs/contributors). Please elaborate.
