### 2022-09-02 #4718 @klb2
> @klb2 May I ask what the second author Eduard A. Jorswieck contributed exactly? He does not seem to made _software contributions_ (see https://github.com/klb2/qmkpy/graphs/contributors). Please elaborate.

Hi @ulf1
Thank you for agreeing to serve as a reviewer of our submission.

It is correct that the second author (Eduard Jorswieck) did not contribute code to the repository. However, he actively contributed behind the scenes to the conception of the framework, he suggested solution algorithms like the implemented round-robin scheme.
In the [documentation of JOSS](https://joss.readthedocs.io/en/latest/submitting.html#authorship) it is stated that
> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are.

Therefore, I consider his co-authorship appropriate due to "active project direction and other forms of non-code contributions".
