### 2022-09-16 #4718 @max-little
Thanks, will be able to respond next week.

On Friday, 16 September 2022, The Open Journals editorial robot <
***@***.***> wrote:
> 👋 @max-little, please update us on how your review is going (this is an
automated reminder).
>
