### 2022-09-17 klb2/qmkpy#2 @ulf1
Multi-line latex formula doesn't work in pypi
Problem:
- No line breaks in latex formula: https://pypi.org/project/qmkpy/

Cause:
- Multiline latex formular in https://github.com/klb2/qmkpy#problem-description
