### 2022-09-17 #4718 @klb2
> Hello @klb2, you can see my current review in [#4718 (comment)](https://github.com/openjournals/joss-reviews/issues/4718#issuecomment-1231505103) and in https://github.com/klb2/qmkpy/issues/created_by/ulf1

Thank you @ulf1 for your thorough review of the QMKPy framework.
I will implement the required changes and close the corresponding issues. Once I am done with everything, I will let you know in this review issue.
