### 2022-09-18 200aed5
Add requirements installation to README; Close #3
Add the pip command to install the Python requirements to the
installation section in the README file.
