### 2022-09-18 e2e422f
Add zip_safe flag to setup; Close #5
Add the zip_safe=True flag to setup to safe some space
