### 2022-09-18 aafc4ff
Add Python version requirement to setup; Close #6
Add the Python version requirement of >=3.7 to the setup
