### 2022-09-18 d4991f5
Add Jupyter notebook with example; Close #9
Add a Jupyter notebook to the examples which illustrates how a novel
algorithm can be easily implemented using qmkpy.
Additionally, a link to binder is provided (both in the README and the
documentation).
