+++ examples/README.md
@@ -13,6 +13,10 @@ All of the examples can simply be run via
 python3 <scriptName>.py
 ```
 
+Additionally, there is a Jupyter notebook provided, which illustrates how a new
+algorithm can be easily implemented.  
+[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/klb2/qmkpy/HEAD?labpath=examples%2FCustom%20Algorithm.ipynb)
+
 
 ## Example List
 The following examples are provided
@@ -23,3 +27,4 @@ The following examples are provided
 | `basic_usage.py` | Simple script where a QMKP is specified and solved. |
 | `custom_algorithm.py` | Illustration of how a custom solution algorithm can be implemented. |
 | `save_problem.py` | Example of how to save a problem instance (using the `txt` strategy) |
+| `Custom Algorithm.ipynb` | Jupyter notebook to illustrate how a custom solution algorithm can be implemented. |
