### 2022-09-20 #4718 @max-little
Message ID: ***@***.***>
--
Max Little (www.maxlittle.net)
Associate Professor, University of Birmingham
Turing Fellow, Alan Turing Institute
TED Fellow (fellows.ted.com/profiles/max-little)
Visiting Associate Professor, MIT
Author: Machine Learning for Signal Processing, Oxford University Press
global.oup.com/academic/product/machine-learning-for-signal-processing-9780198714934
Room 138, School of Computer Science
University of Birmingham
Birmingham B15 2TT
UK +44 7710 609564
Skype: dr.max.little
