### 2022-09-20 #4661 @max-little
On Mon, 29 Aug 2022 at 21:09, The Open Journals editorial robot
***@***.***> wrote:
>
> Closed #4661 as completed.
>
