### 2022-09-20 #4718 @max-little
On Mon, 29 Aug 2022 at 21:08, The Open Journals editorial robot
***@***.***> wrote:
>
> Submitting author: @klb2 (Karl-Ludwig Besser)
> Repository: https://github.com/klb2/qmkpy
> Branch with paper.md (empty if default branch): joss-paper
> Version: v1.1.0
> Editor: @jbytecode
> Reviewers: @ulf1, @max-little
> Archive: Pending
>
> Status
>
> Status badge code:
>
>
>
> Reviewers and authors:
>
> Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the target repository and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)
>
> Reviewer instructions & questions
>
> @ulf1 & @max-little, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
> First of all you need to run this command in a separate comment to create the checklist:
>
> @editorialbot generate my checklist
>
> The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any questions/concerns please let @jbytecode know.
>
> ✨ Please start on your review when you are able, and be sure to complete your review in the next six weeks, at the very latest ✨
>
> Checklists
>
> @ulf1, please create your checklist typing: @editorialbot generate my checklist
>
> @max-little, please create your checklist typing: @editorialbot generate my checklist
>
