### 2022-09-20 #4718 @max-little
Main set of review comments now provided, raised as issues on the project
page.

On Fri, 16 Sept 2022 at 17:33, The Open Journals editorial robot <
***@***.***> wrote:

> 👋 @max-little <https://github.com/max-little>, please update us on how
> your review is going (this is an automated reminder).
>
