### 2022-09-20 #4718 @max-little
- [ ] **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) -> Not quite yet: there needs to be a more extensive review of existing methods for the problem and some implementations of classical methods e.g. dynamic programming, branch-and-bound.

- [x] **Installation:** Does installation proceed as outlined in the documentation? -> I tested this as a manual download, seems to work fine.
- [ ] **Functionality:** Have the functional claims of the software been confirmed? -> To the extent that I was able to run the examples, yes. I have not tried to implement my own solver algorithm, this did not seem to be particularly transparent.

- [ ] **A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is? -> Not really: this needs some work in order to explain better the purpose of this package and why it is needed.
- [ ] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. -> Needs some work to improve the clarity.
- [ ] **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems). -> At least for the pre-existing example algorithms, yes, but not really in terms of explaining how to develop your own algorithms. This needs more careful thought.

- [ ] **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided? -> Not yet, this needs some attention.
- [ ] **A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work? -> Not yet, see above.
- [ ] **State of the field:** Do the authors describe how this software compares to other commonly-used packages? -> Needs work, see above.
- [ ] **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)? -> The references are lacking.
