### 2022-09-21 0ed1ae5
Add module descriptions; Close #16
Add a docstring to each module which documents the purpose of the module
and what type of functions are placed within the module.
