### 2022-09-21 5ee01cf
Add description of variable mapping to docs; Close #14
Add an extensive description of how the mathematical components of the
QMKP (profits, weights, capacities, assignments) are implemented in
qmkpy.
