### 2022-09-21 7b75844
Add remark that profits are aka values; Close #15
A remark is added to the documentation that the profits are also
referred to as "values" in the literature.
