+++ docs/source/index.rst
@@ -34,6 +34,8 @@ weight capacity :math:`c_u`.
 When assigning item :math:`i` to a knapsack, it yields the profit :math:`p_i`.
 When assigning item :math:`j` (with :math:`j\neq i`) to the same knapsack, the
 additional (joint) profit :math:`p_{ij}` is obtained.
+*Remark:* The profits :math:`p` are also referred to as "values" in the
+literature.
 
 A detailed description of the way how the mathematical components of the QMKP
 are implemented in the ``qmkpy`` framework can be found in the `code
