### 2022-09-26 klb2/qmkpy#2 @jbytecode
Due to JOSS submission: https://github.com/openjournals/joss-reviews/issues/4718
