### 2022-10-05 db0a95c
Add numpy version requirement; Close #7
Add a version constraint to the requirements file that numpy version 1
is required.
