+++ requirements.txt
@@ -1,1 +1,1 @@
-numpy
+numpy==1.*
