### 2022-10-05 8d73c9f
Add Contributing section to README; Close #4
Add a reference to the contributing guidelines to the README and add a
link to the documentation to the CONTRIBUTING.md document.
