### 2022-10-05 klb2/qmkpy#2 @klb2
Unfortunately, I was not able to find a solution to this issue so far. The problem appears to be the MathJax version used at pypi. The same formula works fine on GitHub (README) and on ReadTheDocs (index page).
