### 2022-10-05 6859702
Improve mathematical description; Close #12
Improve and fix the mathematical problem description of the QMKP in both
the README and the documentation.
This includes a more precise description of the used variables and
assumptions.
