### 2022-10-05 96450ab
Fix inline math in README
