+++ README.md
@@ -28,12 +28,12 @@ $$
 This describes an assignment problem where one wants to assign $N\in\mathbb{N}$
 items to $K\in\mathbb{N}$ knapsacks, which are described by the index set
 $\mathcal{K}=\{1, 2, \dots, K\}$.
-Item $i$ has the weight $w_i\in\mathbb{R}_{+}$ and knapsack $u$ has the weight
-capacity $c_u\mathbb{R}_{+}$.
+Item $i$ has the weight $w_i\in\mathbb{R_+}$ and knapsack $u$ has the weight
+capacity $c_u\mathbb{R_+}$.
 If item $i$ is assigned to a knapsack, it yields the (non-negative) profit
-$p_i\in\mathbb{R}_{+}$.
+$p_i\in\mathbb{R_+}$.
 If item $j$ (with $j\neq i$ ) is assigned _to the same_ knapsack, we get the
-additional joint profit $p_{ij}\in\mathbb{R}_{+}$.
+additional joint profit $p_{ij}\in\mathbb{R_+}$.
 
 The set of items which are assigned to knapsack $u$ is denoted by
 $\mathcal{A}(u)$ and $a_{iu}\in\{0, 1\}$ is an indicator whether item $i$ is
