### 2022-10-07 #4718 @klb2
I am currently working on fixing the last issues. I will let you know once I am done with this.

However, some issues, e.g. https://github.com/klb2/qmkpy/issues/2, will probably not be fixed. I guess, there needs to be a discussion whether it is necessary to close all of them.
