+++ paper/paper.md
@@ -15,7 +15,7 @@ authors:
 affiliations:
  - name: Institute for Communications Technology, Technische Universität Braunschweig, Germany
    index: 1
-date: 08 August 2022
+date: 07 October 2022
 bibliography: paper.bib
 ---
 
@@ -23,32 +23,37 @@ bibliography: paper.bib
 
 QMKPy provides a Python framework for modeling and solving the quadratic
 multiple knapsack problem (QMKP).
-It is primarily aimed at researchers who develop new solution algorithms for the
-QMKP. 
+It is primarily aimed at researchers who develop new solution algorithms for
+the QMKP.
+QMKPy therefore mostly functions as a testbed to quickly implement novel
+algorithms and compare their results with existing ones.
 However, the package also already includes implementations of established
 algorithms for those who only need to solve a QMKP as part of their research.
 
 The QMKP is a type of knapsack problem which has first been analyzed by
 @Hiley2006.
-For a basic overview of multiple types of knapsack problems see, e.g.,
+For a basic overview of other types of knapsack problems see, e.g.,
 @Kellerer2004.
-As in the regular multiple knapsack problem, the goal is to assign $N$ items
-with given weights $w_i$ and profit values $p_i$ to $K$ knapsacks with given
-weight capacities $c_u$, such that a total profit is maximized.
-In the QMKP, there exist additional joint profits $p_{ij}$ which are yielded
-when items $i$ and $j$ are packed into the same knapsack.
+As in the regular multiple knapsack problem, the goal is to assign
+$N\in\mathbb{N}$ items with given weights $w_i\in\mathbb{R}_{+}$ and
+(non-negative) profit values $p_i\in\mathbb{R}_{+}$ to $K\in\mathbb{N}$
+knapsacks with given weight capacities $c_u\in\mathbb{R}_{+}$, such that a
+total profit is maximized.
+In the QMKP, there exist additional joint profits $p_{ij}\in\mathbb{R}_{+}$
+which are yielded when items $i$ and $j$ are packed into the same knapsack.
 
 Mathematically, the QMKP is described by the following optimization problem
 \begin{subequations}
 \begin{alignat}{3}
 	\max\quad & \sum_{u\in\mathcal{K}}\Bigg(\sum_{i\in\mathcal{A}_u} p_{i} &+&\sum_{\substack{j\in\mathcal{A}_u \\ j\neq i}} p_{ij}\Bigg)\\
 	\mathrm{s.t.}\quad & \sum_{i\in\mathcal{A}_u} w_{i} \leq c_u & \quad & \forall u\in\mathcal{K} \\
-	& \sum_{u=1}^{K} a_{iu} \leq 1  & & \forall 1\leq i \leq N
+	& \sum_{u=1}^{K} a_{iu} \leq 1  & & \forall i\in\{1, 2, \dots{}, N\}
 \end{alignat}
 \end{subequations}
-where $\mathcal{A}_u$ is the set of items that are assigned to knapsack $u$ and
-$a_{iu}\in\{0,1\}$ is a binary variable indicating whether item $i$ is assigned
-to knapsack $u$.
+where $\mathcal{K}=\{1, 2, \dots{}, K\}$ describes the set of knapsacks,
+$\mathcal{A}_u\subseteq\{1, 2, \dots{}, N\}$ is the set of items that are
+assigned to knapsack $u$ and $a_{iu}\in\{0,1\}$ is a binary variable indicating
+whether item $i$ is assigned to knapsack $u$.
 
 
 # Statement of need
@@ -56,33 +61,52 @@ to knapsack $u$.
 The QMKP is a NP-hard optimization problem and therefore, there exists a
 variety of (heuristic) algorithms to find good solutions for it.
 While Python frameworks already exist for the standard (multiple)
[…]
