+++ paper/paper.bib
@@ -32,7 +32,7 @@
 @inproceedings{Besser2022wiopt,
 	author = {Besser, Karl-Ludwig and Jorswieck, Eduard A. and Coon, Justin P.},
 	title = {Multi-User Frequency Assignment for Ultra-Reliable {mmWave} Two-Ray Channels},
-	booktitle = {20th International Symposium on Modeling and Optimization in Mobile, Ad hoc, and Wireless Networks (WiOpt -- RAWNET Workshop)},
+	booktitle = {20th International Symposium on Modeling and Optimization in Mobile, Ad hoc, and Wireless Networks (WiOpt)},
 	year = {2022},
 	month = {9},
 	publisher = {IEEE},
@@ -48,4 +48,37 @@
 	pages = {195--202},
 	publisher = {IOS Press},
 	doi = {10.3233/FAIA200093},
+}
+
+@software{ortools,
+  title = {OR-Tools},
+  version = {v9.4},
+  author = {Laurent Perron and Vincent Furnon},
+  organization = {Google},
+  url = {https://developers.google.com/optimization/},
+  date = {2022-08-11}
+}
+
+@software{QMKPyDatasets,
+  doi = {10.5281/ZENODO.7157144},
+  url = {https://github.com/klb2/qmkpy-datasets},
+  author = {Karl-Ludwig Besser},
+  title = {QMKPy Datasets},
+  version = {v1.0},
+  publisher = {Zenodo},
+  year = {2022},
+  copyright = {Open Access},
+}
+
+@article{Billionnet2004,
+	doi = {10.1287/ijoc.1030.0029},
+	year = {2004},
+	month = {5},
+	publisher = {Institute for Operations Research and the Management Sciences ({INFORMS})},
+	volume = {16},
+	number = {2},
+	pages = {188--197},
+	author = {Alain Billionnet and {\'{E}}ric Soutif},
+	title = {Using a Mixed Integer Programming Tool for Solving the 0--1 Quadratic Knapsack Problem},
+	journal = {INFORMS Journal on Computing}
 }
