### 2022-10-07 3eaa159
Delete PDF generation workflow from master
Delete the GitHub action to generate the JOSS paper draft PDF from the
master branch
