### 2022-10-07 #4718 @jbytecode
@ulf1 - We are waiting our author @klb2 to finalize their edits.

@klb2 - please ping us when your edits are completed.

@ulf1 - I think a special decision has to be made for https://github.com/klb2/qmkpy/issues/2

@max-little - Could you please update your status?
