### 2022-10-07 246f2fd
Apply Black code styling
Format the code files according to the Black Python code styling
