+++ tests/test_qmkp.py
@@ -1,4 +1,4 @@
-#import tempfile
+# import tempfile
 import os.path
 import pathlib
 
@@ -12,43 +12,36 @@ from qmkpy import checks
 
 SAVE_LOAD_STRATEGIES = ("numpy", "pickle", "txt", "json")
 
+
 def test_solver_consistency():
-    profits = np.array([[1, 1, 2, 3],
-                        [1, 1, 4, 5],
-                        [2, 4, 2, 6],
-                        [3, 5, 6, 3]])
+    profits = np.array([[1, 1, 2, 3], [1, 1, 4, 5], [2, 4, 2, 6], [3, 5, 6, 3]])
     weights = [1, 2, 3, 3]
     capacities = [5, 5, 3]
     cp_solution = constructive_procedure(profits, weights, capacities)
-    problem = QMKProblem(profits, weights, capacities,
-                         algorithm=constructive_procedure)
+    problem = QMKProblem(profits, weights, capacities, algorithm=constructive_procedure)
     pr_solution, profit = problem.solve()
     assert np.all(pr_solution == cp_solution) and profit > 0
 
+
 def test_parameter_consistency():
-    profits = np.array([[1, 1, 2, 3],
-                        [1, 1, 4, 5],
-                        [2, 4, 2, 6],
-                        [3, 5, 6, 3]])
+    profits = np.array([[1, 1, 2, 3], [1, 1, 4, 5], [2, 4, 2, 6], [3, 5, 6, 3]])
     weights = [1, 2, 3, 3]
     capacities = [5, 5, 3]
     cp_solution = constructive_procedure(profits, weights, capacities)
-    problem = QMKProblem(profits, weights, capacities,
-                         algorithm=constructive_procedure)
+    problem = QMKProblem(profits, weights, capacities, algorithm=constructive_procedure)
     pr_solution, profit = problem.solve()
     assert np.all(problem.capacities == [5, 5, 3])
 
+
 def test_parameter_writeable():
-    profits = np.array([[1, 1, 2, 3],
-                        [1, 1, 4, 5],
-                        [2, 4, 2, 6],
-                        [3, 5, 6, 3]])
+    profits = np.array([[1, 1, 2, 3], [1, 1, 4, 5], [2, 4, 2, 6], [3, 5, 6, 3]])
     weights = [1, 2, 3, 3]
     capacities = [5, 5, 3]
     problem = QMKProblem(profits, weights, capacities)
     with pytest.raises(ValueError):
         problem.capacities[0] = 2
     assert np.all(problem.capacities == [5, 5, 3])
+
 
 def test_solver_set_later():
     num_elements = 20
@@ -63,6 +56,7 @@ def test_solver_set_later():
     pr_solution, profit = problem.solve()
     assert np.all(pr_solution == cp_solution) and profit > 0
 
+
 def test_solver_with_args():
     num_elements = 20
     num_knapsacks = 5
@@ -70,11 +64,12 @@ def test_solver_with_args():
     profits = profits @ profits.T
     weights = np.random.randint(1, 5, size=(num_elements,))
     capacities = np.random.randint(3, 12, size=(num_knapsacks,))
-    problem = QMKProblem(profits, weights, capacities,
-                         algorithm=fcs_procedure,
-                         args=(.5, 10))
+    problem = QMKProblem(
+        profits, weights, capacities, algorithm=fcs_procedure, args=(0.5, 10)
+    )
     pr_solution, profit = problem.solve()
[…]
