### 2022-10-07 520e1be
Fix style issues; Close #8
Fix minor style issues like unassigned variables in both main source
code files and unit tests.
