### 2022-10-07 #4718 @klb2
I have now closed all remaining issues, except for https://github.com/klb2/qmkpy/issues/2.

Please let me know, if there is anything else to do.
