### 2022-10-11 #4718 @max-little
@klb2 You can find a fairly in-depth set of review comments raised as
issues on the project page. Let me know if you need any
clarifications.

On Thu, 8 Sept 2022 at 11:27, Mehmet Hakan Satman
***@***.***> wrote:
>
> @ulf1, @max-little - could you please update us on how is going your review?
>
> @max-little - you haven't generated your task list yet. could you please generate?
>
> Thank you in advance.
>
