### 2022-10-11 #4718 @klb2
I have implemented the requested changes and closed the issues raised on the project page by both reviewers @ulf1 and @max-little .

The updated paper can be downloaded from https://github.com/openjournals/joss-reviews/issues/4718#issuecomment-1271515159
