### 2022-10-11 #4718 @max-little
OK thanks. Yes, I'll take a look at these this week.
Best
Max

On Tue, 11 Oct 2022 at 10:16, Mehmet Hakan Satman ***@***.***>
wrote:

> @editorialbot <https://github.com/editorialbot> generate pdf
>
> @max-little <https://github.com/max-little> - Could you please check the
> corrections applied by our author.
>
> Thank you in advance.
>
