### 2022-10-12 #4718 @jbytecode
@klb2 - You decide to use $R_{+}$ instead of $R^{+}$ in the manuscript. Is that why superscripting is reserved for representing dimension, e.g., $R_{+}^{2}$? Is that the usual notation in the corresponding literature? Thank you in advance.
