### 2022-10-12 #4718 @klb2
> @klb2 - You decide to use R+ instead of R+ in the manuscript. Is that why superscripting is reserved for representing dimension, e.g., R+2? Is that the usual notation in the corresponding literature? Thank you in advance.
>
> Edit:
>
> R>0=Set(x∈R|x>0)=R+
>
> is the usual definition.

I am aware that different notations exist for the same set. As you mentioned, we typically use the index $R_{+}$ to indicate the non-negativity and the superscript to indicate the dimension, e.g., $\mathbb{R}_{+}^{2}$.
