### 2022-10-23 klb2/qmkpy#18 @max-little
 Very "thin" review of the area e.g. missing exact solver examples? #17
"While many solution algorithms are proposed in the literature, the primary goal of the framework is to provide a testbed environment for researchers and developers working in that area to quickly implement and share their novel solutions."

OK, so, I think the appropriate solution would be to make this very clear in the title of the package: i.e. it is not a package providing solutions for the quadratic multiple knapsack problem, it is a package providing a testbed for the development and testing of quadratic multiple knapsack algorithms. Otherwise the package name and title are quite misleading.
