### 2022-10-23 #4718 @max-little
Sorry for the late reply. I believe I am basically satisfied with the
responses here, apart from the naming of the package; I've raised this in
https://github.com/klb2/qmkpy/issues/18

Max

On Sun, 23 Oct 2022 at 07:09, Mehmet Hakan Satman ***@***.***>
wrote:

> Dear @max-little <https://github.com/max-little>
>
> Could you please update your status and inform us? Do we have any progress?
>
> Thank you in advance and have a good weekend.
>
