### 2022-10-24 #4718 @max-little
Yes, that sounds good!
M.

On Monday, 24 October 2022, Karl-Ludwig Besser ***@***.***>
wrote:
> Thank you for your comments @max-little
>
> Would you agree that "QMKPy: A Python Testbed for the Quadratic Multiple
Knapsack Problem" is a more appropriate title for the package/paper?
> This should better highlight the nature of the package (that it primarily
serves as a testbed for algorithms to solve the QMKP).
>
