### 2022-10-24 7312745
Change title to "Testbed"; Close #18
Change the (sub)title of the package and the paper to "Testbed" (from
framework) to highlight the nature of the QMKPy package.
