### 2022-10-24 #4718 @klb2
Great, thank you @max-little.
I have changed the title accordingly in both the README and the paper.

@jbytecode: Is there any way/need of changing the title of the JOSS submission? I guess, you would have to do this. The new title is "QMKPy: A Python Testbed for the Quadratic Multiple Knapsack Problem". Thank you.
