### 2022-10-25 #4718 @klb2
I have read the paper, created a tagged release, and the Zenodo archive.

- The version is `v1.2.0`.
- The Zenodo DOI is 10.5281/zenodo.7249626
- The Zenodo archive URL is https://zenodo.org/record/7249626
