### 2022-10-31 #4718 @gkthiruvathukal
@klb2  I've just done a read through of the paper. It appears to be in great shape overall. One of the references seems to be lacking a DOI:

```latex
@inproceedings{Besser2022wiopt,
	author = {Besser, Karl-Ludwig and Jorswieck, Eduard A. and Coon, Justin P.},
	title = {Multi-User Frequency Assignment for Ultra-Reliable {mmWave} Two-Ray Channels},
	booktitle = {20th International Symposium on Modeling and Optimization in Mobile, Ad hoc, and Wireless Networks (WiOpt)},
	year = {2022},
	month = {9},
	publisher = {IEEE},
}
```

Is there any chance you could fix this? Oddly, it doesn't show up as a missing DOI in the report above.
