### 2022-11-01 #4718 @jbytecode
@gkthiruvathukal - As far as I know, the editorialbot does not perform a DOI check for `inproceedings` type bibtex entries. That's why we didn't see a missing DOI warning.
