### 2022-11-02 #4718 @gkthiruvathukal
@klb2 Yes, even I found that a bit strange (the DOI check issue). I think I'm going to proceed with the final steps.
