### 2022-11-02 #4718 @klb2
Thank you very much @gkthiruvathukal for finalizing the paper.

Thank you @jbytecode @ulf1 @max-little for handling and reviewing the paper/software.

It was a really interesting and fun way to organize the review process.
